refactor(account): use dynamic import for subcommand
